requirements.txt
asgiref==3.2.7
attrs==19.3.0
awsebcli==3.18.1
bcrypt==3.1.7
blessed==1.17.5
boto3==1.14.4
botocore==1.17.4
cached-property==1.5.1
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.6
django-cors-headers==3.3.0
django-scheduler==0.9.3
django-sslserver==0.22
django-storages==1.9.1
django-tailwind==0.9.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.9
importlib-metadata==1.6.0
jmespath==0.10.0
jsonschema==3.2.0
oauthlib==3.1.0
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.0
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
urllib3==1.25.9
watchtower==0.7.3
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.1.0